+++ gradle-baseline-java/src/main/groovy/com/palantir/baseline/plugins/javaversions/BaselineJavaVersions.java
@@ -54,8 +54,8 @@ public final class BaselineJavaVersions implements Plugin<Project> {
             throw new GradleException(String.format(
                     "BaselineJavaVersions requires %s. %s is not supported", MIN_GRADLE_VERSION, currentGradleVersion));
         }
-        RootBaselineJavaVersionsExtension rootExtension =
-                project.getExtensions().create(EXTENSION_NAME, RootBaselineJavaVersionsExtension.class, project);
+        BaselineJavaVersionsExtension rootExtension =
+                project.getExtensions().create(EXTENSION_NAME, BaselineJavaVersionsExtension.class, project);
         project.subprojects(proj ->
                 proj.getExtensions().create(EXTENSION_NAME, SubprojectBaselineJavaVersionsExtension.class, proj));
