+++ gradle-baseline-java/src/main/groovy/com/palantir/baseline/plugins/javaversions/SubprojectBaselineJavaVersionsExtension.java
@@ -20,7 +20,7 @@ import javax.inject.Inject;
 import org.gradle.api.GradleException;
 import org.gradle.api.Project;
 
-public class SubprojectBaselineJavaVersionsExtension implements BaselineJavaVersionsExtension {
+public class SubprojectBaselineJavaVersionsExtension implements BaselineJavaVersionsExtensionSetters {
     private final Project project;
 
     @Inject
